sched/tune: Only disable the CPU margin instead of completely disabling schedtune
It is better that tasks, which are either boosted or are under prefer idle, are placed accordingly to the cores they're meant to be at always once configured. This is to prevent massive migration once schedtune interactive replenishes. Thus, remove per-task checks and only retain the checks for the CPU boost.

Signed-off-by: ederekun <sedrickvince@gmail.com>