sched/tune: Filter tasks being boosted according to their adj value
Inspired by Yaroslav Furman's patch, with this approach; instead of comparing the name of the cgroup for every task, we instead store the index of the concerning schedtune group during boostgroup init and use that instead to see whether the task falls within the group we want to filter.
We can define multiple groups this way without noticeably slowing down the schedtune_boost function. In addition to this, according to empirical testings, having set the adj value to 100 instead of 0 allows sufficient CPU boosting and enough important tasks to be boosted to bring noticeable improvement to system responsiveness as compared to not having the filter - taking into consideration Yaroslav's implementation utilizing Sultan's CPU input boost of which we don't have.

Signed-off-by: ederekun <sedrickvince@gmail.com>